(refactor) Made modular backend routers and controllers, all endpoints working
