+++ dateServer/routers/routes.js
@@ -1,7 +1,9 @@
-var routeController = require('../routes/site');
+var routeController = require('../controllers/site');
+var userController = require('../controllers/users');
 
 module.exports = function (app) {
 
   app.get('/', routeController.index);
+  app.post('/bestmatch', userController.getMatchingEvents);
 
 }
